# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: Front Office API — league management, results, standings, and scheduling.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues/{leagueId}/dq-findings:
    get:
      operationId: listDqFindings
      summary: List data-quality findings for a league (commissioner only)
      tags: [DQ]
      parameters:
        - $ref: '#/components/parameters/LeagueId'
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [BLOCK, ALERT, WATCH]
        - name: resolved
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated list of DQ findings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DqFindingsEnvelope'
        '403':
          description: Not the league commissioner.
        '404':
          description: League not found.
  /leagues/{leagueId}/dq-findings/{findingId}/resolve:
    patch:
      operationId: resolveDqFinding
      summary: Mark a DQ finding as resolved (commissioner only)
      tags: [DQ]
      parameters:
        - $ref: '#/components/parameters/LeagueId'
        - name: findingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Finding marked resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DqFinding'
        '403':
          description: Not the league commissioner.
        '404':
          description: Finding not found.
...
components:
  parameters:
...
    LeagueId:
      name: leagueId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    DqFinding:
      type: object
      required: [id, checked_at, check_name, severity, detail]
      properties:
        id:
          type: integer
        checked_at:
          type: string
          format: date-time
        check_name:
          type: string
        severity:
          type: string
          enum: [BLOCK, ALERT, WATCH]
        league_id:
          type: ['string', 'null']
        season_id:
          type: ['string', 'null']
        entity_id:
          type: ['string', 'null']
        detail:
          type: string
          description: JSON string — the full row from the check view, serialised with row_to_json().
        resolved_at:
          type: ['string', 'null']
          format: date-time
        resolved_by:
          type: ['string', 'null']
    DqFindingsEnvelope:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DqFinding'
        total:
          type: integer

    # ------------------------------------------------------------------ Competition (existing)
...